Move result_from_quota to PeriodicLimiter
In writing the documentation I realized that the behaviour of
retry_after in that case (and in what would have been a more general
case) didn't make any sense. I moved it to the PeriodicLimiter where the
logic mainly fits and updated it so that it works the way I would expect
it to with retry_after.