Allow stores to provide the clock for limiters
This fixes an oversight that I forgot about before releasing 2018.12.0
that is especially important for our GCRA limiter. The stores can
ostensibly retrieve the time from their backing object (redis, some
other data store) and that time will be more likely to be consistent
than the local time across multiple machines.

This updates the version and release notes as well so we can release
this.