Add an example Flask project
Let's add some example projects to Rush as well so people can see how it
might be used.

This is a single endpoint that checks for auth credentials and throttles
based on those. It doesn't attempt to verify the credentials and I don't
care about that.

Other folks can expand on this, but this shows exactly what I wanted to
show, so I'm happy.