Implement CAS for RedisStore
This commit changes the behaviour of the RedisStore from using hashes
to store LimitData objects to JSON strings at individual keys. This is
because Redis is unable to WATCH hashes, meaning that it is not possible
to perform an atomic compare and swap operation without aquiring a lock
which could have significant performance implications.