Fix last of the linter errors
So the redis-py typeshed stubs are wrong and I intend to fix them. Also,
trying to handle async and sync code in the same decorator seems to be a
_really_ bad idea but this is contrib code so I'm not that bothered by
it.

Finally, let's prepare for `python setup.py check -r -s` to go away and
start using `twine check` but let's make it easy to lint by using a
temporary directory intended specifically for it.